BaseService: return UInt8 for method_re1b

+++ YogaSMC/YogaBaseService.hpp
@@ -123,7 +123,7 @@ protected:
      *
      *  @return kIOReturnSuccess on success
      */
-    IOReturn method_re1b(UInt32 offset, UInt32 *result);
+    IOReturn method_re1b(UInt32 offset, UInt8 *result);
 
     /**
      *  Wrapper for RECB
